requirements.txt
python-dotenv
ccxt